Bundled settlement-assign
